BacklogApiProcessing/requirements.txt
azure-functions==1.0.5
certifi==2018.10.15
chardet==3.0.4
python-dateutil==2.8.1
idna==2.7
jmespath==0.9.3
pip==19.0.3
pybacklog==0.1.7
pytz==2018.7
PyYAML==5.1b5
requests==2.20.1
ruamel.appconfig==0.5.4
ruamel.std.argparse==0.8.1
setuptools==39.0.1
urllib3==1.24.2
zope.interface==4.6.0
